requirements.txt
pandas>=2.0.0
xlrd>=2.0.1
openpyxl>=3.1.0
rapidfuzz>=3.0.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
